initial casfinder support

+++ lib/Bio/Adventure/Annotation.pm
@@ -25,6 +25,80 @@ use List::MoreUtils qw"any";
 use Template;
 use Text::CSV_XS::TSV;
 
+=head2 C<Casfinder>
+
+Invocation of casfinder, it currently is using the singularity image
+because for reasons I have not yet figured out, it does not appear to
+work otherwise.
+
+=cut
+sub Casfinder {
+    my ($class, %args) = @_;
+    my $options = $class->Get_Vars(
+        args => \%args,
+        required => ['input'],
+        jprefix => '22',
+        jmem => 8,
+        cpus => 4,
+        modules => ['casfinder'],);
+    my $loaded = $class->Module_Loader(modules => $options->{modules});
+    my $check = which('casfinder.sh');
+    die('Could not find casfinder in your PATH.') unless($check);
+
+    my $job_name = $class->Get_Job_Name();
+    my $inputs = $class->Get_Paths($options->{input});
+    my $cwd_name = basename(cwd());
+
+    my $casfinder_exe_dir = dirname($check);
+    ## Hey, don't forget abs_path requires a file which already exists.
+    my $input_filename = basename($options->{input});
+    my $output_filename = qq"${input_filename}.tsv";
+    my $input_dir = dirname($options->{input});
+    my $input_dirname = basename($input_dir);
+    my $input_path = abs_path($input_dir);
+    $input_path = qq"${input_path}/${input_filename}";
+    my $output_dir = qq"outputs/$options->{jprefix}casfinder_${input_dirname}";
+    my $comment = qq!## This is a casfinder submission script
+!;
+    my $stdout = qq"${output_dir}/casfinder.stdout";
+    my $stderr = qq"${output_dir}/casfinder.stderr";
+    my $jstring = qq!mkdir -p ${output_dir}
+start=\$(pwd)
+cd ${output_dir}
+casfinder.sh $options->{input} \\
+  2>casfinder.stderr \\
+  1>casfinder.stdout
+test=\$?
+if [[ "\${test}" -eq "0" ]]; then
+  echo "Casfinder succeeded."
+else
+  echo "Casfinder failed."
+  exit \$?
+fi
+cd \${start}
+!;
+    my $casfinder = $class->Submit(
+        cpus => 8,
+        comment => $comment,
+        jdepends => $options->{jdepends},
+        jname => "casfinder_${job_name}",
+        jprefix => $options->{jprefix},
+        jstring => $jstring,
+        jmem => 16,
+        modules => $options->{modules},
+        output => qq"${output_dir}/casfinder.tsv",
+        prescript => $options->{prescript},
+        postscript => $options->{postscript},
+        jqueue => 'large',
+        stdout => $stdout,
+        stderr => $stderr,
+        walltime => '144:00:00',);
+
+    $loaded = $class->Module_Loader(modules => $options->{modules},
+                                    action => 'unload');
+    return($casfinder);
+}
+
 =head2 C<Interproscan>
 
  Invoke interproscan on a set of sequences.
@@ -83,11 +157,17 @@ perl -pe 's/\\*//g' ${input_path} > ${input_filename}
 interproscan.sh -i ${input_filename} \\
   2>interproscan.stderr \\
   1>interproscan.stdout
+test=\$?
+if [[ "\${test}" -eq "0" ]]; then
+  rm -rf temp
+else
+  echo "Interproscan failed."
+  exit \$?
+fi
 if [[ -f interproscan.tsv ]]; then
   rm interproscan.tsv
 fi
 ln -sf "${output_filename}" interproscan.tsv
-rm -rf temp
 cd \${start}
 !;
     my $interproscan = $class->Submit(
